+++ .storybook/manager.ts
@@ -1,6 +1,6 @@
 // .storybook/manager.js
 
-import { addons } from '@storybook/addons';
+import { addons } from '@storybook/manager-api';
 
 addons.setConfig({
     isFullscreen: false,
