+++ .storybook/preview.ts
@@ -1,0 +1,15 @@
+import type { Preview } from '@storybook/svelte';
+import './global.css';
+
+const preview: Preview = {
+	parameters: {
+		controls: {
+			matchers: {
+				color: /(background|color)$/i,
+				date: /Date$/i
+			}
+		}
+	}
+};
+
+export default preview;
