+++ package.json
@@ -31,19 +31,21 @@
   },
   "devDependencies": {
     "@babel/core": "^7.23.9",
-    "@storybook/addon-actions": "7.6.10",
-    "@storybook/addon-essentials": "7.6.10",
-    "@storybook/addon-interactions": "7.6.10",
-    "@storybook/addon-links": "7.6.10",
-    "@storybook/addon-mdx-gfm": "7.6.10",
+    "@chromatic-com/storybook": "1.2.18",
+    "@storybook/addon-actions": "^8.0.0",
+    "@storybook/addon-essentials": "^8.0.0",
+    "@storybook/addon-interactions": "^8.0.0",
+    "@storybook/addon-links": "^8.0.0",
+    "@storybook/blocks": "^8.0.0",
     "@storybook/addon-styling": "^1.3.7",
-    "@storybook/addon-svelte-csf": "^4.1.0",
-    "@storybook/svelte": "7.6.10",
-    "@storybook/sveltekit": "7.6.10",
-    "@storybook/testing-library": "^0.2.2",
+    "@storybook/addon-svelte-csf": "^4.1.2",
+    "@storybook/svelte": "^8.0.0",
+    "@storybook/sveltekit": "^8.0.0",
+    "@storybook/test": "^8.0.0",
     "@sveltejs/adapter-auto": "3.1.1",
     "@sveltejs/kit": "2.5.0",
     "@sveltejs/package": "2.1.0",
+    "@sveltejs/vite-plugin-svelte": "^3.0.2",
     "@types/d3": "^7.4.3",
     "@typescript-eslint/eslint-plugin": "^6.19.1",
     "@typescript-eslint/parser": "^6.19.1",
@@ -52,24 +54,23 @@
     "date-fns": "^2.30.0",
     "eslint": "^8.56.0",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-storybook": "^0.6.15",
+    "eslint-plugin-storybook": "^0.8.0",
     "eslint-plugin-svelte3": "^4.0.0",
     "lodash": "^4.17.21",
     "prettier": "^3.2.4",
     "prettier-plugin-svelte": "^3.1.2",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "storybook": "7.6.10",
+    "storybook": "^8.0.0",
     "svelte-check": "^3.6.3",
     "svelte-loader": "^3.1.9",
     "svelte-preprocess": "^5.1.3",
     "svelte-tools": "^1.2.6",
     "tslib": "^2.6.2",
     "typescript": "^5.3.3",
-    "vite": "^5.0.12"
+    "vite": "^5.1.4"
   },
   "dependencies": {
     "@floating-ui/dom": "^1.6.3",
+    "@storybook/test": "8.0.0",
     "d3": "^7.8.5",
     "hash-it": "^6.0.0",
     "svelte": "^4.2.9",
