+++ src/stories/Introduction.mdx
@@ -1,4 +1,5 @@
-import { Meta } from '@storybook/addon-docs';
+import { Meta } from "@storybook/blocks";
+
 import Code from './assets/code-brackets.svg';
 import Colors from './assets/colors.svg';
 import Comments from './assets/comments.svg';
